# docs/swagger.yaml
basePath: /
definitions:
...
  models.Correo:
    properties:
      area:
        type: string
      created_at:
        type: string
      direccion:
        type: string
      dueno:
        $ref: '#/definitions/models.Persona'
      dueno_id:
        type: string
      id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
...
  models.Direccion:
    properties:
      area:
        type: string
      created_at:
        type: string
      dueno:
        $ref: '#/definitions/models.Persona'
      dueno_id:
        type: string
      empleados:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      estado:
        type: string
      id:
        type: string
      lugar:
        type: string
      municipio:
        type: string
      nombre:
        type: string
      parroquia:
        type: string
      updated_at:
        type: string
      urbanizacion:
        type: string
      userID:
        type: string
      usuarios:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
    type: object
...
  models.Empresa:
    properties:
      actividad_economica:
        type: string
      area:
        type: string
      created_at:
        type: string
      direccion:
        type: string
      empleados:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      id:
        type: string
      nombre:
        type: string
      representante:
        $ref: '#/definitions/models.Persona'
      representante_id:
        type: string
      rif:
        type: string
      socios:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      updated_at:
        type: string
      userID:
        type: string
    type: object
...
  models.IIO:
    properties:
      area:
        type: string
      created_at:
        type: string
      descripcion:
        type: string
      fecha:
        type: string
      id:
        type: string
      imagen_url:
        type: string
      lugar:
        type: string
      modalidad:
        type: string
      nivel:
        type: string
      nombre:
        type: string
      parroquia:
        type: string
      redi:
        type: string
      relacion:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tie:
        type: string
      updated_at:
        type: string
      urbanizacion:
        type: string
      userID:
        type: string
      verificada:
        type: boolean
      zodi:
        type: string
    type: object
...
  models.Nacionalidad:
    properties:
      area:
        type: string
      created_at:
        type: string
      gentilicio:
        type: string
      id:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
  models.Pasaporte:
    properties:
      area:
        type: string
      codigo:
        type: string
      created_at:
        type: string
      foto:
        type: string
      id:
        type: string
      numero:
        type: string
      pais:
        type: string
      representante:
        $ref: '#/definitions/models.Persona'
      representante_id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
  models.Persona:
    properties:
      alias:
        type: string
      apellido:
        type: string
      area:
        type: string
      cargo_actual:
        type: string
      cedula:
        type: string
      correo:
        type: string
      correos:
        items:
          $ref: '#/definitions/models.Correo'
        type: array
      created_at:
        type: string
      direcciones:
        items:
          $ref: '#/definitions/models.Direccion'
        type: array
      empresas:
        items:
          $ref: '#/definitions/models.Empresa'
        type: array
      estado_civil:
        type: string
      filiacion_politica:
        type: string
      id:
        type: string
      ideologia:
        type: string
      iio:
        items:
          $ref: '#/definitions/models.IIO'
        type: array
      informacion_de_interes:
        type: string
      nacionalidad:
        items:
          $ref: '#/definitions/models.Nacionalidad'
        type: array
      nombre:
        type: string
      pasaporte:
        items:
          $ref: '#/definitions/models.Pasaporte'
        type: array
      profesion:
        type: string
      redes:
        items:
          $ref: '#/definitions/models.Redes'
        type: array
      relacionados:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      religion:
        type: string
      telefono:
        type: string
      tipo_perfil:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      valoraciones:
        type: string
      vehiculos:
        items:
          $ref: '#/definitions/models.Vehiculo'
        type: array
      visa:
        items:
          $ref: '#/definitions/models.Visa'
        type: array
    type: object
  models.Redes:
    properties:
      area:
        type: string
      created_at:
        type: string
      direccion:
        type: string
      dueno:
        $ref: '#/definitions/models.Persona'
      dueno_id:
        type: string
      id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
...
  models.Vehiculo:
    properties:
      area:
        type: string
      año:
        type: string
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      linea:
        type: string
      marca:
        type: string
      matricula:
        type: string
      modelo:
        type: string
      numero:
        type: string
      propietario:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      usuarios:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      vehiculo:
        items:
          $ref: '#/definitions/models.IIO'
        type: array
    type: object
  models.Visa:
    properties:
      aprobada:
        type: boolean
      area:
        type: string
      codigo:
        type: string
      created_at:
        type: string
      id:
        type: string
      pais:
        type: string
      representante:
        $ref: '#/definitions/models.Persona'
      representante_id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      valoracion:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API SISTEMA DE EXPLORACION POPULAR.
  title: API
  version: "1.0"
paths:
...
  /correos:
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos de la dirección a crear
        in: body
        name: correo
        required: true
        schema:
          $ref: '#/definitions/models.Correo'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tipo del correo
        in: body
        name: tipo
        required: true
        schema:
          type: string
      - description: Área del correo
        in: body
        name: area
        required: true
        schema:
          type: string
      - description: Dirección del correo
        in: body
        name: direccion
        required: true
        schema:
          type: string
      - description: ID del representante
        in: body
        name: representante_id
        required: true
        schema:
          type: string
      - description: ID del usuario
        in: body
        name: user_id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Correo'
      summary: Crea un nuevo correo
      tags:
      - Correo
  /correos/{id}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del correo a eliminar
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Correo'
        "204":
          description: Correo eliminado exitosamente
      summary: Elimina un correo por su ID
      tags:
      - Correo
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del correo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Correo'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Correo'
      summary: Obtiene un correo por su ID
      tags:
      - Correo
    put:
      consumes:
      - application/json
      parameters:
      - description: Datos del correo a actualizar
        in: body
        name: direccion
        required: true
        schema:
          $ref: '#/definitions/models.Correo'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del correo a actualizar
        in: path
        name: id
        required: true
        type: string
      - description: Tipo del correo
        in: body
        name: tipo
        required: true
        schema:
          type: string
      - description: Área del correo
        in: body
        name: area
        required: true
        schema:
          type: string
      - description: Dirección del correo
        in: body
        name: direccion
        required: true
        schema:
          type: string
      - description: ID del representante
        in: body
        name: representante_id
        required: true
        schema:
          type: string
      - description: ID del usuario
        in: body
        name: user_id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Correo'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Correo'
      summary: Actualiza un correo existente por su ID
      tags:
      - Correo
  /direcciones:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Datos de la dirección a crear
        in: body
        name: direccion
        required: true
        schema:
          $ref: '#/definitions/models.Direccion'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Direccion'
      summary: Crea una nueva dirección
      tags:
      - Direccion
  /direcciones/{id}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID de la dirección a eliminar
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Dirección eliminada exitosamente
      summary: Elimina una dirección por su ID
      tags:
      - Direccion
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID de la dirección
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Direccion'
      summary: Obtiene una dirección por su ID
      tags:
      - Direccion
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID de la dirección a actualizar
        in: path
        name: id
        required: true
        type: string
      - description: Datos de la dirección a actualizar
        in: body
        name: direccion
        required: true
        schema:
          $ref: '#/definitions/models.Direccion'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Direccion'
      summary: Actualiza una dirección existente por su ID
      tags:
      - Direccion
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"